Fix VisuallyHidden to export props + move it to core
