+++ src/index.tsx
@@ -76,5 +76,8 @@ export {
 export { Paragraph, ParagraphProps } from './core/Paragraph/Paragraph';
 export { Text, TextProps } from './core/Text/Text';
 export { Textarea, TextareaProps } from './core/Form/Textarea/Textarea';
-export { VisuallyHidden } from './core/VisuallyHidden/VisuallyHidden';
+export {
+  VisuallyHidden,
+  VisuallyHiddenProps,
+} from './core/VisuallyHidden/VisuallyHidden';
 export { suomifiDesignTokens, DesignTokens } from 'suomifi-design-tokens';
